+++ frontend/src/components/features/station/search/StationSearchForm.tsx
@@ -9,10 +9,10 @@ import { Button } from '~/components/ui/button'
 import { useMidpointMutation } from '~/hooks/useMidpointMutation'
 import { logger } from '~/lib/logger'
 import { stationSearchSchema } from '~/schemas/station-search.schema'
-import StationAutocomplete from '../autocomplete/StationAutocomplete'
-import { AddFormButton } from '../buttons/AddFormButton'
-import { RemoveFormButton } from '../buttons/RemoveFormButton'
-import { ResetFormButton } from '../buttons/ResetFormButton'
+import StationAutocomplete from './StationAutocomplete'
+import { AddFormButton } from '../../../ui/buttons/AddFormButton'
+import { RemoveFormButton } from '../../../ui/buttons/RemoveFormButton'
+import { ResetFormButton } from '../../../ui/buttons/ResetFormButton'
 import { Form, FormControl, FormField, FormItem } from './Form'
 
 const MAX_AREA_FIELDS = 6
