+++ frontend/src/components/features/station/search/buttons/RemoveFormButton.tsx
